@@ -161,9 +161,9 @@
     "jsdom": "^20.0.0",
     "plist": "^3.0.5",
     "qrcode": "^1.5.1",
-    "react": "17",
     "react-blockies": "^1.4.1",
-    "react-dom": "17",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-i18next": "^11.16.7",
     "react-router-dom": "^6.3.0",
     "recharts": "^2.1.14",
@@ -215,15 +215,15 @@
     "@types/jest": "^27.4.1",
     "@types/node": "^17.0.25",
     "@types/plist": "^3.0.2",
-    "@types/react": "^18.0.6",
-    "@types/react-dom": "^18.0.2",
+    "@types/react": "^18.0.24",
+    "@types/react-dom": "^18.0.8",
     "@types/react-router-dom": "^5.3.3",
     "@types/supertest": "^2.0.12",
     "@types/tmp": "^0.2.3",
     "@typescript-eslint/eslint-plugin": "^5.20.0",
     "@typescript-eslint/parser": "^5.20.0",
-    "@vitejs/plugin-react": "^2.0.0",
-    "electron": "^21.1.1",
+    "@vitejs/plugin-react": "^2.2.0",
+    "electron": "^21.2.1",
     "electron-builder": "^23.1.0",
     "electron-devtools-installer": "^3.2.0",
     "eslint": "^8.13.0",
@@ -239,9 +239,9 @@
     "supertest": "^6.2.4",
     "tmp": "^0.2.1",
     "ts-jest": "^28.0.7",
-    "typescript": "^4.6.3",
-    "vite": "^3.0.3",
-    "vite-plugin-electron": "^0.8.3",
-    "vite-plugin-svgr": "^2.2.1"
+    "typescript": "^4.8.4",
+    "vite": "^3.2.2",
+    "vite-plugin-electron": "^0.10.2",
+    "vite-plugin-svgr": "^2.2.2"
   }
 }
