@@ -68,7 +68,8 @@ const fallBackImage = 'fallback'
 const epicLoginUrl = 'https://legendary.gl/epiclogin'
 const gogLoginUrl =
   'https://auth.gog.com/auth?client_id=46899977096215655&redirect_uri=https%3A%2F%2Fembed.gog.com%2Fon_login_success%3Forigin%3Dclient&response_type=code&layout=galaxy'
-const sidInfoUrl = 'https://github.com/G7DAO/HyperPlay/issues/42'
+const sidInfoUrl =
+  'https://github.com/Heroic-Games-Launcher/HeroicGamesLauncher/wiki/How-To:-Epic-Alternative-Login'
 const githubURL = 'https://github.com/G7DAO/HyperPlay/releases/latest'
 const GITHUB_API = 'https://api.github.com/repos/G7DAO/HyperPlay/releases'
 const supportURL = 'https://github.com/G7DAO/HyperPlay/blob/main/Support.md'
